Make user_switch into a macro
