+++ myia/macros.py
@@ -4,12 +4,21 @@ A macro transforms a subgraph into another. It is run during inference, which
 means it has access to type information.
 """
 
+from functools import reduce
 
 from . import abstract, operations
-from .abstract import macro
-from .dtype import Array, Number
+from .abstract import ANYTHING, TYPE, VALUE, macro, type_token, union_simplify
+from .dtype import Array, Bool, Number
 from .info import About, DebugInfo
-from .ir import Constant, Graph, MetaGraph, MultitypeGraph, Parameter
+from .ir import (
+    CloneRemapper,
+    Constant,
+    Graph,
+    GraphCloner,
+    MetaGraph,
+    MultitypeGraph,
+    Parameter,
+)
 from .prim import ops as P
 from .prim.py_implementations import scalar_cast, scalar_to_array, typeof
 from .utils import (
@@ -196,6 +205,156 @@ async def dict_values(info):
     getters = [info.graph.apply(P.dict_getitem, dref.node, k)
                for k in typ.entries]
     return info.graph.apply(P.make_tuple, *getters)
+
+
+class _CastRemapper(CloneRemapper):
+
+    def __init__(self,
+                 graphs,
+                 inlines,
+                 manager,
+                 relation,
+                 graph_relation,
+                 clone_constants,
+                 graph_repl,
+                 fv_replacements):
+        """Initialize the GraphCloner."""
+        super().__init__(
+            graphs=graphs,
+            inlines=inlines,
+            manager=manager,
+            relation=relation,
+            graph_repl=graph_repl,
+            graph_relation=graph_relation,
+            clone_constants=clone_constants,
+        )
+        self.fv_replacements = fv_replacements
+
+    def gen_fv(self, g, ng, fv):
+        """Remap the free variables we want to remap."""
+        if fv in self.fv_replacements:
+            new = self.fv_replacements[fv]
+            self.remap_node((g, fv), g, fv, ng, new, link=False)
+
+
+@macro
+async def user_switch(info):
+    """Implement the switch functionality generated by the parser.
+
+    If user_switch finds a Union in the condition, it will infer the value of
+    the condition for each type in the union. If the condition is necessarily
+    true or false for some types, the type of the variable for the
+    corresponding conditional branch will be set to these types.
+    """
+    engine = info.engine
+    g = info.graph
+    condref, tbref, fbref = check_nargs(P.switch, 3, info.argrefs)
+
+    async def type_trials(focus, opnode, argrefs):
+        """Handle `user_switch(hastype(x, typ), tb, fb)`.
+
+        We want to evaluate tb in a context where x has type typ and fb
+        in a context where it doesn't.
+        """
+        def cond_trial(cg, opt):
+            # For each possible type we make a "cond trial" which replaces the
+            # focus input in the condition function by one that's cast to the
+            # type. We can thus check if the value of the condition depends
+            # directly on the type.
+            return cg.apply(opnode,
+                            *nodes[:focus],
+                            cg.apply(P.unsafe_static_cast, nodes[focus], opt),
+                            *nodes[focus + 1:])
+
+        async def wrap(branch_ref, branch_type):
+            # We transform branch_graph into a new graph which refers to a cast
+            # version of x. We also transform all of the children of x's graph
+            # so that closures called in the branch also refer to the cast
+            # version of x.
+            branch_graph = branch_ref.node.value
+            if branch_graph not in xg.scope:
+                return branch_graph
+            rval = branch_graph.make_new(relation='copy')
+            cast = rval.apply(P.unsafe_static_cast, xref.node, branch_type)
+            cl = GraphCloner(
+                *xg.children,
+                total=False,
+                graph_repl={branch_graph: rval},
+                remapper_class=_CastRemapper.partial(
+                    fv_replacements={xref.node: cast}
+                )
+            )
+            assert rval is cl[branch_graph]
+            engine.mng.add_graph(rval)
+            return rval
+
+        nodes = [ref.node for ref in argrefs]
+        xref = argrefs[focus]
+        fulltype = await xref.get()
+        assert isinstance(fulltype, abstract.AbstractUnion)
+
+        xg = xref.node.graph
+        cg = condref.node.graph
+        cond_trials = [cond_trial(cg, t) for t in fulltype.options]
+        results = [await engine.ref(node, condref.context).get()
+                   for node in cond_trials]
+
+        groups = {True: [], False: [], ANYTHING: []}
+
+        for t, result in zip(fulltype.options, results):
+            assert isinstance(result, abstract.AbstractScalar)
+            assert result.values[TYPE] is Bool
+            value = result.values[VALUE]
+            groups[value].append(t)
+
+        if groups[ANYTHING]:
+            return await default()
+
+        tbtyp = union_simplify(groups[True])
+        fbtyp = union_simplify(groups[False])
+
+        if tbtyp is None:
+            return fbref
+        elif fbtyp is None:
+            return tbref
+        else:
+            new_conds = [g.apply(P.hastype, xref.node, t)
+                         for t in groups[True]]
+            new_cond = reduce(lambda x, y: g.apply(P.bool_or, x, y),
+                              new_conds)
+            new_tb = await wrap(tbref, tbtyp)
+            new_fb = await wrap(fbref, fbtyp)
+            return g.apply(P.switch, new_cond, new_tb, new_fb)
+
+    async def default():
+        _, cond, tb, fb = info.outref.node.inputs
+        condt = await condref.get()
+        if not engine.check_predicate(Bool, type_token(condt)):
+            to_bool = engine.pipeline.resources.convert(bool)
+            cond = g.apply(to_bool, cond)
+        return g.apply(P.switch, cond, tb, fb)
+
+    for branch_ref in [tbref, fbref]:
+        if not branch_ref.node.is_constant_graph():
+            raise MyiaTypeError(
+                'Branches of switch must be functions when the condition'
+                ' is hastype on a Union.'
+            )
+
+    ctx = condref.context
+    if condref.node.is_apply():
+        opnode, *args = condref.node.inputs
+        opref = engine.ref(opnode, ctx)
+        ops = (await opref.get()).get_sync()
+        if len(ops) == 1:
+            op, = ops
+            argrefs = [engine.ref(a, ctx) for a in args]
+            argtypes = [await arg.get() for arg in argrefs]
+            for i, arg in enumerate(argtypes):
+                if isinstance(arg, abstract.AbstractUnion):
+                    return await type_trials(i, opnode, argrefs)
+
+    return await default()
 
 
 @macro
