+++ myia/pipeline/resources.py
@@ -54,7 +54,7 @@ scalar_object_map = {
     operations.next: C.next,
     operations.to_array: C.to_array,
     operations.switch: P.switch,
-    operations.user_switch: P.user_switch,
+    operations.user_switch: M.user_switch,
     operations.slice: Slice,
     operations.apply: M.apply,
     math.floor: P.scalar_floor,
@@ -110,7 +110,7 @@ standard_object_map = {
     operations.next: C.next,
     operations.to_array: C.to_array,
     operations.switch: P.switch,
-    operations.user_switch: P.user_switch,
+    operations.user_switch: M.user_switch,
     operations.slice: Slice,
     operations.apply: M.apply,
     math.floor: P.scalar_floor,
